demos/rag-visual-simulator/requirements.txt
streamlit>=1.32.0
openai>=1.12.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
